fix: 중복 드래그 이슈 해결
