+++ packages/web/src/components/index.ts
@@ -5,3 +5,4 @@ export { default as QueryProvider } from './QueryProvider';
 export { default as ThemeProvider } from './ThemeProvider';
 export { default as ToggleTheme } from './ToggleTheme';
 export { default as YoutubeEmbed } from './YoutubeEmbed';
+export { default as DragBox } from './DragBox';
